修改权限结构

+++ src/pages/dataLoad/list/list.tsx
@@ -22,6 +22,7 @@ import noDataElement from '@/components/no-data';
 import modal from '@/pages/workflowConfig/tools/edit-custom-collection-modal/modal';
 import { PermissionWrapper } from '@/components/PermissionGuard';
 import { DATA_LOAD_PERMISSIONS } from '@/config/permissions';
+import { OperationColumn } from '@ccf2e/arco-material';
 export enum RunState {
   SUCCEED = 'succeed',
   FAILED = 'failed',
@@ -270,36 +271,31 @@ export default function DataLoad() {
       fixed: 'right',
       width: 105,
       render: (_, item) => {
+        const perms = item?.perms || [];
+        const config = [] as any;
+        if (perms.includes(DATA_LOAD_PERMISSIONS.CAN_GET)) {
+          config.push({
+            label: '详情',
+            onClick: () => {
+              gotoDetail(item.task_id);
+            }
+          });
+        }
+        if (perms.includes(DATA_LOAD_PERMISSIONS.CAN_DETELE)) {
+          config.push({
+            label: '删除',
+            onClick: () => {
+              deleteLoad(item.task_id, item.name);
+            }
+          });
+        }
         return (
-          <div
-            className={Styles.hoverStyle}
-            style={{
-              width: '100%',
-              display: 'flex',
-              justifyContent: 'space-around'
-            }}
-          >
-            {item.perms.includes(DATA_LOAD_PERMISSIONS.CAN_GET) && (
-              <span
-                className={Styles.hoverStyle}
-                onClick={() => {
-                  gotoDetail(item.task_id);
-                }}
-              >
-                详情
-              </span>
-            )}
-            {item.perms.includes(DATA_LOAD_PERMISSIONS.CAN_DETELE) && (
-              <span
-                className={Styles.hoverStyle}
-                onClick={() => {
-                  deleteLoad(item.task_id, item.name);
-                }}
-              >
-                删除
-              </span>
-            )}
-          </div>
+          <OperationColumn
+            row={item}
+            config={config}
+            index={0}
+            extendFont="操作"
+          />
         );
       }
     }
